requirements.txt
# Core Django
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
djongo==1.3.6
pymongo[srv]==3.12.3

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1

# Environment variables
python-dotenv==1.0.1

# API Documentation
drf-yasg==1.21.7

# Image Processing
Pillow==10.2.0

# AI/ML Integration
openai==1.12.0
numpy==1.26.4
pandas==2.2.0

# Security
django-ratelimit==4.1.0
django-filter==23.5

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.0.1
pytest-django==4.8.0
coverage==7.4.1

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
